Removed teacher endpoints and changed load grades to be only for students
