+++ src/main/java/ch/teachu/teachuapi/grade/GradeRepo.java
@@ -1,25 +1,23 @@
 package ch.teachu.teachuapi.grade;
 
-import ch.teachu.teachuapi.errorhandling.NotFoundException;
 import ch.teachu.teachuapi.generated.tables.User;
-import ch.teachu.teachuapi.generated.tables.records.GradeRecord;
-import ch.teachu.teachuapi.grade.dto.*;
+import ch.teachu.teachuapi.grade.dto.GradeResponse;
+import ch.teachu.teachuapi.grade.dto.SchoolClassGradesResponse;
+import ch.teachu.teachuapi.grade.dto.SemesterGradesResponse;
+import ch.teachu.teachuapi.grade.dto.SubjectGradesResponse;
 import ch.teachu.teachuapi.parent.AbstractRepo;
 import ch.teachu.teachuapi.util.DateUtil;
-import ch.teachu.teachuapi.util.DoubleHolder;
-import org.jooq.Condition;
+import ch.teachu.teachuapi.util.Holder;
 import org.springframework.stereotype.Repository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 
 import static ch.teachu.teachuapi.generated.tables.Exam.EXAM;
 import static ch.teachu.teachuapi.generated.tables.Grade.GRADE;
 import static ch.teachu.teachuapi.generated.tables.SchoolClass.SCHOOL_CLASS;
 import static ch.teachu.teachuapi.generated.tables.SchoolClassSubject.SCHOOL_CLASS_SUBJECT;
-import static ch.teachu.teachuapi.generated.tables.SchoolClassUser.SCHOOL_CLASS_USER;
 import static ch.teachu.teachuapi.generated.tables.Semester.SEMESTER;
 import static ch.teachu.teachuapi.generated.tables.Subject.SUBJECT;
 import static ch.teachu.teachuapi.generated.tables.User.USER;
@@ -31,62 +29,92 @@ public class GradeRepo extends AbstractRepo {
     private static final String SUBJECT_TEACHER = "subject_teacher";
     private static final String STUDENT = "student";
 
-    public UUID createGrade(CreateGradeRequest request, UUID studentId) {
-        UUID id = UUID.randomUUID();
-        GradeRecord grade = sql().newRecord(GRADE);
-        grade.setId(id);
-        grade.setStudentId(studentId);
-        grade.setMark(request.getMark());
-        grade.setNote(request.getNote());
-        grade.setExamId(request.getExamId());
-        grade.store();
-        return id;
-    }
+    public List<SemesterGradesResponse> loadGrades(UUID studentId) {
+        List<SemesterGradesResponse> semesters = new ArrayList<>();
 
-    public boolean studentHasGrade(UUID studentId, UUID examId) {
-        return sql().fetchExists(sql().select().from(GRADE)
-                .where(GRADE.STUDENT_ID.eq(studentId), GRADE.EXAM_ID.eq(examId)));
-    }
+        Holder<SemesterGradesResponse> currentSemester = new Holder<>();
+        Holder<SchoolClassGradesResponse> currentSchoolClass = new Holder<>();
+        Holder<SubjectGradesResponse> currentSubject = new Holder<>();
 
-    public boolean studentCanTakeExam(UUID studentId, UUID examId) {
-        return sql().fetchExists(sql().select().from(EXAM)
-                .join(SCHOOL_CLASS_SUBJECT).on(EXAM.SCHOOL_CLASS_SUBJECT_ID.eq(SCHOOL_CLASS_SUBJECT.ID))
-                .join(SCHOOL_CLASS_USER).on(SCHOOL_CLASS_SUBJECT.SCHOOL_CLASS_ID.eq(SCHOOL_CLASS_USER.SCHOOL_CLASS_ID))
-                .where(EXAM.ID.eq(examId), SCHOOL_CLASS_USER.USER_ID.eq(studentId)));
-    }
+        User classTeacher = USER.as(CLASS_TEACHER);
+        User subjectTeacher = USER.as(SUBJECT_TEACHER);
+        User student = USER.as(STUDENT);
 
-    public void changeGrade(ChangeGradeRequest request, UUID studentId) {
-        GradeRecord grade = sql().selectFrom(GRADE).where(GRADE.ID.eq(request.getId())).fetchOptional()
-                .orElseThrow(() -> new NotFoundException("Grade"));
-        grade.setStudentId(studentId);
-        grade.setMark(request.getMark());
-        grade.setNote(request.getNote());
-        grade.setExamId(request.getExamId());
-        grade.store();
-    }
-
-    public GradesResponse loadWithRestriction(LoadGradeRequest request) {
-        List<Condition> conditions = new ArrayList<>();
-        addCondition(conditions, EXAM.ID.eq(request.getExamId()), request.getExamId());
-        addCondition(conditions, USER.ID.eq(request.getStudentId()), request.getStudentId());
-
-        DoubleHolder sum = new DoubleHolder();
-        DoubleHolder weightSum = new DoubleHolder();
-        List<ExamGradeResponse> examGrades = sql().select(GRADE.ID, USER.FIRST_NAME, USER.LAST_NAME, GRADE.MARK, GRADE.NOTE, EXAM.WEIGHT,
-                        EXAM.NAME, EXAM.DESCRIPTION, EXAM.DATE, EXAM.VIEW_DATE)
+        // load all grades in one stream and sort it
+        sql().select(SEMESTER.ID,
+                        SEMESTER.NAME,
+                        SCHOOL_CLASS.ID,
+                        SCHOOL_CLASS.NAME,
+                        classTeacher.FIRST_NAME,
+                        classTeacher.LAST_NAME,
+                        SUBJECT.ID,
+                        SUBJECT.NAME,
+                        SUBJECT.WEIGHT,
+                        subjectTeacher.FIRST_NAME,
+                        subjectTeacher.LAST_NAME,
+                        GRADE.MARK,
+                        GRADE.NOTE,
+                        EXAM.ID,
+                        EXAM.NAME,
+                        EXAM.DESCRIPTION,
+                        EXAM.WEIGHT,
+                        EXAM.DATE,
+                        EXAM.VIEW_DATE,
+                        student.FIRST_NAME,
+                        student.LAST_NAME,
+                        student.ID)
                 .from(GRADE)
-                .join(USER).on(GRADE.STUDENT_ID.eq(USER.ID))
+                .join(student).on(GRADE.STUDENT_ID.eq(student.ID))
                 .join(EXAM).on(GRADE.EXAM_ID.eq(EXAM.ID))
-                .where(conditions)
-                .stream().map(record -> {
-                    sum.plusSet(record.value6() * record.value4());
-                    weightSum.plusSet(record.value6());
-                    return new ExamGradeResponse(record.value1(), record.value7(), record.value8(), record.value6(),
-                            DateUtil.toDate(record.value9()), DateUtil.toDate(record.value10()), record.value4(),
-                            record.value5(), record.value2(), record.value3());
-                })
-                .collect(Collectors.toList());
+                .join(SCHOOL_CLASS_SUBJECT).on(EXAM.SCHOOL_CLASS_SUBJECT_ID.eq(SCHOOL_CLASS_SUBJECT.ID))
+                .join(subjectTeacher).on(SCHOOL_CLASS_SUBJECT.TEACHER_ID.eq(subjectTeacher.ID))
+                .rightJoin(SUBJECT).on(SCHOOL_CLASS_SUBJECT.SUBJECT_ID.eq(SUBJECT.ID))
+                .join(SCHOOL_CLASS).on(SCHOOL_CLASS_SUBJECT.SCHOOL_CLASS_ID.eq(SCHOOL_CLASS.ID))
+                .join(classTeacher).on(SCHOOL_CLASS.TEACHER_ID.eq(classTeacher.ID))
+                .join(SEMESTER).on(EXAM.SEMESTER_ID.eq(SEMESTER.ID))
+                .where(student.ID.eq(studentId))
+                .orderBy(SEMESTER.FROM.desc(), SCHOOL_CLASS.NAME, SUBJECT.NAME, EXAM.DATE.desc())
+                .stream().forEach(record -> {
+                    // create the grade itself
+                     GradeResponse grade = new GradeResponse(record.value15(), record.value16(), record.value17(),
+                            DateUtil.toDate(record.value18()), DateUtil.toDate(record.value19()),
+                             record.value12(), record.value13());
 
-        return new GradesResponse(sum.divide(weightSum), examGrades);
+                    // check if it is a new semester
+                    if (!currentSemester.exists() || !currentSemester.get().getSemesterId().equals(record.get(0))) {
+                        SemesterGradesResponse semesterExamsResponse = new SemesterGradesResponse(record.value1(),
+                                record.value2(), new ArrayList<>());
+
+                        semesters.add(semesterExamsResponse);
+                        currentSemester.set(semesterExamsResponse);
+                        // force the child objects to create a new object
+                        currentSchoolClass.clear();
+                        currentSubject.clear();
+                    }
+
+                    // same as semester for school class
+                    if (!currentSchoolClass.exists() || !currentSchoolClass.get().getSchoolClassId().equals(record.get(2))) {
+                        SchoolClassGradesResponse schoolClassExamsResponse = new SchoolClassGradesResponse(record.value3(),
+                                record.value4(), record.value5(), record.value6(), 0, new ArrayList<>());
+
+                        currentSchoolClass.set(schoolClassExamsResponse);
+                        currentSemester.get().getSchoolClasses().add(schoolClassExamsResponse);
+                        currentSubject.clear();
+                    }
+
+                    // same as semester for subject
+                    if (!currentSubject.exists() || !currentSubject.get().getSubjectId().equals(record.get(6))) {
+                        SubjectGradesResponse subjectExamsResponse = new SubjectGradesResponse(record.value7(),
+                                record.value8(), record.value10(), record.value11(),
+                                record.value9(), 0, new ArrayList<>());
+
+                        currentSubject.set(subjectExamsResponse);
+                        currentSchoolClass.get().getSubjects().add(subjectExamsResponse);
+                    }
+
+                    currentSubject.get().getGrades().add(grade);
+                });
+
+        return semesters;
     }
 }
