+++ src/main/java/ch/teachu/teachuapi/grade/dto/GradeResponse.java
@@ -1,4 +1,4 @@
-package ch.teachu.teachuapi.exam.dto;
+package ch.teachu.teachuapi.grade.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,17 +6,17 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import java.util.Date;
-import java.util.UUID;
 
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class ChangeExamRequest {
-    private UUID id;
+public class GradeResponse {
     private String name;
     private String description;
     private double weight;
     private Date date;
     private Date viewDate;
+    private double mark;
+    private String note;
 }
